% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_weather.R
\name{actualvd}
\alias{actualvd}
\title{Calculate Actual Vapor Density}
\usage{
actualvd(Ta, h, ...)
}
\arguments{
\item{Ta}{vector of air temperature (◦C).}

\item{h}{vector of relative humidity (\%).}

\item{...}{additional parameters accepted by function \code{satvp}.}
}
\value{
vector of actual vapor density (kg m-3).
}
\description{
calculates actual vapor density (kg m-3). Only valid for temperature above
0◦C.
}
\examples{
data("p1_tah")
temperature <- p1_tah$Ta
humidity <- p1_tah$h
actualvd(temperature, humidity)
}
