% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_gwp.R
\name{AGWP.CH4}
\alias{AGWP.CH4}
\title{AGWP of CH4}
\usage{
AGWP.CH4(TH = 100, mRF_CH4 = 107.3805)
}
\arguments{
\item{TH}{chosen time horizon (i.e. 20, 100, 500 years)}

\item{mRF_CH4}{marginal radiative forcing of methane (W kg-1). It may change
between different years, so please keep up the pace of IPCC report.}
}
\value{
a scalar
}
\description{
Calculates absolute global warming potential (W m-2 kg-1 yr) of methane for a given period.
}
\examples{
##1 AGWP CH4 for 100 years
AGWP.CH4()

##2 AGWP CH4 for 20 years
AGWP.CH4(TH = 20)

##3 relative GWP CH4 for 100 years
AGWP.CH4()/AGWP.CO2()
}
