% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_gwp.R
\name{AGWP.CO2}
\alias{AGWP.CO2}
\title{AGWP of CO2}
\usage{
AGWP.CO2(TH = 100, mRF_CO2 = 0.908)
}
\arguments{
\item{TH}{chosen time horizon (i.e. 20, 100, 500 years)}

\item{mRF_CO2}{marginal radiative forcing of carbon dioxide (W kg-1). It may change
between different years, so please keep up the pace of IPCC report.}
}
\value{
a scalar
}
\description{
Calculates absolute global warming potential (W m-2 kg-1 yr) of carbon dioxide for a given period.
}
\examples{
##1 AGWP CO2 for 100 years
AGWP.CO2()

##2 AGWP CO2 for 20 years
AGWP.CO2(TH = 20)
}
