% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_gwp.R
\name{AGWP.N2O}
\alias{AGWP.N2O}
\title{AGWP of N2O}
\usage{
AGWP.N2O(TH = 100, mRF_N2O = 185.1244)
}
\arguments{
\item{TH}{chosen time horizon (i.e. 20, 100, 500 years)}

\item{mRF_N2O}{marginal radiative forcing of nitrogen oxide (W kg-1). It may change
between different years, so please keep up the pace of IPCC report.}
}
\value{
a scalar
}
\description{
Calculates absolute global warming potential (W m-2 kg-1 yr) of nitrogen oxide for a given period.
}
\examples{
##1 AGWP N2O for 100 years
AGWP.N2O()

##2 AGWP N2O for 20 years
AGWP.N2O(TH = 20)

##3 relative GWP N2O for 100 years
AGWP.N2O()/AGWP.CO2()
}
