% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_weather.R
\name{dewT}
\alias{dewT}
\title{calculates Dew Point Temperature}
\usage{
dewT(Ta, h, ...)
}
\arguments{
\item{Ta}{vector of air temperature (◦C).}

\item{h}{vector of relative humidity (\%).}

\item{...}{additional parameters accepted by function \code{satvp}.}
}
\value{
vector of dew point temperature (◦C).
}
\description{
calculates dew point temperature (◦C). Only valid for temperature above
0◦C.
}
\examples{
data("p1_tah")
temperature <- p1_tah$Ta
humidity <- p1_tah$h
dewT(temperature, humidity)
}
\references{
Fritschen, L. J., and Gay, L. W. (1979). Environmental Instrumentation.
Springer Science & Business Media. 216pp
}
