% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_weather.R
\name{dh}
\alias{dh}
\title{Calculate Daylight Hours}
\usage{
dh(lati, doy)
}
\arguments{
\item{lati}{latitude of the site in degree.}

\item{doy}{day of the year. It can be calculated by the function "\code{DOY}".}
}
\value{
value(s) of daylight hours.
}
\description{
calculates daylight hours. It is the maximum possible during sunshine or daylight hours (h).
}
\examples{
doy <- DOY(2022,9,3)
lati <- -20
dh(lati, doy)
}
\references{
Allen, R. G., Pereira, L. S., Raes, D. and Smith, M. (1998). Crop Evapotranspiration – Guidelines for Computing Crop Water Requirements, FAO Irrigation and Drainge Paper 56, FAO, 1998, ISBN 92-5-104219-5.
}
