% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_weather.R
\name{DOY}
\alias{DOY}
\title{Calculate DOY}
\usage{
DOY(year, month, day)
}
\arguments{
\item{year}{vector of years.}

\item{month}{vector of months.}

\item{day}{vector of days.}
}
\value{
vector of DOY (day of the years).
}
\description{
calculates day of the year.
}
\examples{
## return a day number of September 22nd in 2016
DOY(2016,9,22)

## return multiple day numbers
years <- c(2016,2017,2018)
months <- c(9,10,2)
days <- c(22,13,14)
DOY(years, months, days)
}
