% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_gwp.R
\name{GWP.albedo}
\alias{GWP.albedo}
\title{GWP From Albedo Change}
\usage{
GWP.albedo(
  Sw_up,
  Sw_down,
  alpha_s_REF,
  Ta,
  N = 1,
  TH = 100,
  mRF_CO2 = 0.908,
  method = "IRF"
)
}
\arguments{
\item{Sw_up}{reflected light (W m-2).}

\item{Sw_down}{total incident sunlight (W m-2).}

\item{alpha_s_REF}{the albedo value of the reference site.}

\item{Ta}{upwelling transmittance value. It can be calculated by function \code{transmittance}.}

\item{N}{the number of days. Default 1.}

\item{TH}{chosen time horizon (i.e. 20, 100, 500 years). Default 100.}

\item{mRF_CO2}{marginal radiative forcing of carbon dioxide (W kg-1). It may change
between different years, so please keep up the pace of IPCC report.}

\item{method}{either "\code{IRF}” or \code{"AGWP"} method.}
}
\value{
a value of relative global warming potential
}
\description{
Calculates relative global warming potential due to albedo change.
}
\examples{
alpha_REF <- 83/482
GWP.albedo(134, 478, alpha_REF, 0.854)
}
\references{
Chen, J. (2021). Biophysical Models and Applications in Ecosystem Analysis.
Higher Education Press. 138pp
}
