% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_weather.R
\name{heatflux}
\alias{heatflux}
\title{Calculate Heat Flux Density}
\usage{
heatflux(detT, d, kappa = 2.5)
}
\arguments{
\item{detT}{temperature difference (◦C/K) between the top and the bottom of the soil layer.}

\item{d}{thickness of the soil layer (m).}

\item{kappa}{the thermal conductivity of the soil (W m−1 K−1). It is ~ 2.5 (W m−1 K−1)
for soil minerals, ~1.92 (W m−1 K−1) for organic matter, and 4.18 (W m−1 K−1) for water.}
}
\value{
a vector of heat flux density.
Chen, J. (2021). Biophysical Models and Applications in Ecosystem Analysis.
Higher Education Press. 37pp
}
\description{
calculates heat flux density (W m-2) of soil.
}
\examples{
## soil, delta T = 2, d = 1m
heatflux(2, 1)

## soil minerals, delta T = 2, d = 1m
heatflux(2, 1, kappa = 2.5)

## organic matter, delta T = 2, d = 1m
heatflux(2, 1, kappa = 1.92)

## water, delta T = 2, d = 1m
heatflux(2, 1, kappa = 4.18)
}
