% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_weather.R
\name{heatstore}
\alias{heatstore}
\title{Calculate Heat Storage}
\usage{
heatstore(detT, dettime, theta, roub = 2.7)
}
\arguments{
\item{detT}{temperature difference (◦C/K) over a period of time.}

\item{dettime}{period of time (s).}

\item{theta}{the volumetric soil water content (\%).}

\item{roub}{the soil bulk density (kg m−1).}
}
\value{
a vector of heat storage.
}
\description{
calculates the heat storage (W m-3) in soil over a period of time.
}
\examples{
## delta T = 3 Celsius/Kelvin, 6 hours, and soil water content of 50\%
heatstore(3, 3600*6, 50)
}
\references{
Chen, J. (2021). Biophysical Models and Applications in Ecosystem Analysis.
Higher Education Press. 37pp
}
