% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_weather.R
\name{hourlyT}
\alias{hourlyT}
\title{Calculate Hourly Air Temperature}
\usage{
hourlyT(
  hours,
  t_max,
  t_min,
  time_sr,
  time_ss,
  alpha = 1.5,
  beta = 4,
  gamma = 1
)
}
\arguments{
\item{hours}{time in hours (from 1 to 24) used for calculating air temperature}

\item{t_max}{daily maximum air temperature (Celsius or Fahrenheit)}

\item{t_min}{daily minimum temperature (Celsius or Fahrenheit)}

\item{time_sr}{sunrise time (hour)}

\item{time_ss}{sunset time (hour)}

\item{alpha}{time lag for maximum air temperature (hour). Default 1.5.}

\item{beta}{nighttime temperature coefficient. Default 4.}

\item{gamma}{time lag for minimum air temperature (hour). Default 1.}
}
\value{
a vector of predicted air temperature
}
\description{
calculates hourly air temperature based on the algorithm proposed by
Parton and Logan (1981).
}
\examples{
lng <- 85.446758
lat <- 42.476708
t_max <- 30.2
t_min <- 14.3
doy <- DOY(2016, 9, 22)
hours <- seq(0, 23.5, 0.5)            # half-hour time resolution
time_sr <- sunRS(doy, lng, lat)[[1]]  # calculate sunrise time
time_ss <- sunRS(doy, lng, lat)[[2]]  # calculate sunset time
hourlyT(hours, t_max, t_min, time_sr, time_ss,
        alpha = 1.5, beta = 4, gamma = 1)

}
\references{
Parton, W. J., and Logan, J. A. (1981). A model for diurnal variation in
soil and air temperature. Agricultural Meteorology, 23, 205–216.\cr
}
