% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_evapotranspiration.R
\name{m1.PET.FAO}
\alias{m1.PET.FAO}
\title{Calculate PET using Penman-Monteith Model}
\usage{
m1.PET.FAO(ta, vpd, Rn, u2, G, EL, Cn = 900, Cd = 0.34)
}
\arguments{
\item{ta}{vector of air temperature (◦C).}

\item{vpd}{vapor pressure deficient (kPa).}

\item{Rn}{vector of net radiation (MJ m-2)}

\item{u2}{vector of wind speed (m s-1) at 2 m height.}

\item{G}{vector of soil heat flux (MJ m−2).}

\item{EL}{elevation (m).}

\item{Cn}{numerator constant thta changes with reference surface and
calculation time step (900 ◦C mm s−3 Mg−1 d−1 for 24 h time steps, and
37 ◦C mm s−3 Mg−1 d−1 for hourly time steps).}

\item{Cd}{denominator constant that changes with reference surface and
calculation steps (0.34 s m−1 for 24 h time steps, 0.24 s m-1 for hourly
time steps during daytime, and 0.96 s m-1 for hourly nighttime for grass
reference surface).}
}
\value{
vector of potential evapotranspiration (mm d-1).
}
\description{
Calculates potential evapotranspiration (to be more exactly,
grass reference ET, mm). Only valid for temperature above 0◦C. In
additional, this model assumes a stand that has a 0.12 m canopy height,
a leaf area index (LAI) of 4.8, a bulk surface resistence of 70 s m-1,
and an albedo of 0.23.
}
\examples{
ta <- c(21.17, 23.17, 23.65, 19.66)
vpd <- c(1.15, 0.95, 0.84, 0.67)
Rn <- c(163.70, 174.69, 185.77, 125.35)*24*3600/1e6 #convert from W m-2 to MJ m-2 d-1
u2 <- c(2.04, 1.79, 1.52, 1.75)
G <- c(-5.52, -6.64, -8.13, -1.17)*24*3600/1e6 #convert from W m-2 to MJ m-2 d-1
EL <- 0
m1.PET.FAO(ta, vpd, Rn, u2, G, EL)
}
\references{
Allen, R. G., Smith, M., Perrier, A., and Pereira, L. S. (1994). An update for the definition of reference evapotranspiration. International Commission on Irrigation and Drainage Bulletin, 43(2), 1–34.\cr
Djaman, K., Koudahe, K., Lombard, K., and O’Neill, M. (2018). Sum of hourly vs daily Penman–Monteith grass-reference evapotranspiration under semiarid and arid climate. Irrigation and Drainage Engineering, 7(1), doi: 10.4172/2168-9768.1000202.\cr
Chen, J. (2021). Biophysical Models and Applications in Ecosystem Analysis.
Higher Education Press. 98pp
}
