% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_evapotranspiration.R
\name{m2.PET.Thornthwaite}
\alias{m2.PET.Thornthwaite}
\title{Calculate PET using Thornthwaite Model}
\usage{
m2.PET.Thornthwaite(month_ta, Ld)
}
\arguments{
\item{month_ta}{a vector of 12 elements, each showing the mean air temperature for a month within a year.}

\item{Ld}{a vector of 12 elements, each showing the mean daytime length (12·h), it is time from sunrise to sunset in multiples of 12 hours.}
}
\value{
vector of monthly potential evapotranspiration (cm).
}
\description{
Calculates monthly potential evapotranspiration (cm) based on method
proposed by Thornthwaite (1948), the most widely used temperature-based monthly
scale PET model.
}
\examples{
ta <- c(1.05, 4.53, 9.31, 12.01, 19.99, 25.17, 27.17, 27.65, 23.66, 17.25, 11.61, 2.39)
doy <- DOY(2016, 9, 22)
daylength <- diff(sunRS(doy, 85.44675833, 42.47670833))[[1]]
daylength <- daylength/12 #hours to multiples of 12 hours
m2.PET.Thornthwaite(ta, daylength)
}
\references{
Thornthwaite, C. W. (1948). An approach toward a Rational classification of climate. Geographical Review, 38(1), 55–94.\cr
Chen, J. (2021). Biophysical Models and Applications in Ecosystem Analysis.
Higher Education Press. 99pp
}
