% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_evapotranspiration.R
\name{m3.ET.Fang}
\alias{m3.ET.Fang}
\title{Calculate ET (Method 3)}
\usage{
m3.ET.Fang(PET, VPD, Rn)
}
\arguments{
\item{PET}{monthly potential evapotranspiration (mm month-1) estimated by Hamon's method (function \code{"m3.PET.Hamon"}).}

\item{VPD}{vapor pressure deficit (kPa)}

\item{Rn}{solar radiation (MJ m-2 month-1)}
}
\value{
vector of monthly evapotranspiration (mm month-1)
}
\description{
Calculates evapotranspiration (mm month−1) using the model proposed by Fang et al. (2016)
}
\examples{
PET <- c(-2.26, 70.5, 204.07, 246.24, 386.44, 424.31, 452.8, 481.52, 324.91, 172.19, 86.08, 0.57)
VPD <- c(0.15, 0.17, 0.34, 0.55, 0.91, 1.15, 0.95, 0.84, 0.67, 0.40, 0.32, 0.11)
Rn <- c(-0.87, 27.20, 78.73, 95.00, 149.09, 163.70, 174.69, 185.77, 125.35, 66.43, 33.21, 0.22)
m3.ET.Fang(PET, VPD, Rn)
}
\references{
Fang, Y., Sun, G., Caldwell, P., McNulty, S. G., Noormets, A., Domec, J. C., King, J., Zhang, Z., Zhang, X., Lin, G., Zhou, G., Xiao, J., Chen, J., and Zhou, G. (2016). Monthly land cover-specific evapotranspiration models derived from global eddy flux measurements and remote sensing data. Ecohydrology, 9(2), 248–266.
Chen, J. (2021). Biophysical Models and Applications in Ecosystem Analysis.
Higher Education Press. 104pp
}
