% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_evapotranspiration.R
\name{m3.PET.Hamon}
\alias{m3.PET.Hamon}
\title{Calculate PET using Hamon's model}
\usage{
m3.PET.Hamon(ta, Ld)
}
\arguments{
\item{ta}{mean air temperature (◦C)}

\item{Ld}{mean daytime length (12·h), it is time from sunrise to sunset in multiples of 12 hours.}
}
\value{
vector of daily potential evapotranspiration (mm).
}
\description{
Calculates daily potential evapotranspiration (mm) based on Hamon's PET
model.
}
\examples{
ta <- c(-3.05, -1.53, 5.31, 8.01, 15.99, 21.17, 23.17, 23.65, 19.66, 13.25, 7.61, -2.39)
doy <- DOY(2016, 9, 22)
daylength <- diff(sunRS(doy, 85.44675833, 42.47670833))[[1]]
daylength <- daylength/12 #hours to multiples of 12 hours
m3.PET.Hamon(ta, daylength)
}
\references{
Hamon, W. R. (1963). Computation of direct runoff amounts from storm rainfall. International Association of Scientific Hydrology Publication, 63, 52–62.\cr
Chen, J. (2021). Biophysical Models and Applications in Ecosystem Analysis.
Higher Education Press. 100pp
}
