% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_evapotranspiration.R
\name{m4.ET.type1}
\alias{m4.ET.type1}
\title{Calculate ET (Method 4)}
\usage{
m4.ET.type1(
  Rn,
  PET = NULL,
  LAI = NULL,
  P = NULL,
  SWC = NULL,
  Ta = NULL,
  land.type = c("shrub", "crop", "grass", "deciduous", "needle", "broad", "mixed",
    "savanna")
)
}
\arguments{
\item{Rn}{monthly net solar radiation (MJ month-1)}

\item{PET}{potential evapotranspiration (mm month-1) estimated by Hamon's method (function \code{"m3.PET.Hamon"}).}

\item{LAI}{leaf area index}

\item{P}{predication (mm month-1)}

\item{SWC}{soil water content (\%)}

\item{Ta}{mean air temperature (◦C)}

\item{land.type}{type of land. "\code{shrub}" for shrubland,
"\code{crop}" for cropland, "\code{grass}" for grassland,
"\code{deciduous}" for deciduous forest, "\code{needle}" for evergreen needle leaf forest,
"\code{broad}" for evengreen broad leaf forest, "\code{mixed}" for mixed forest, "\code{savannas}" for savannas.}
}
\value{
vector of monthly evapotranspiration (mm month-1)
}
\description{
Calculates evapotranspiration (mm month−1) using the model proposed by Fang et al. (2016)
}
\examples{
Rn <- c(-0.87, 27.20, 78.73, 95.00, 149.09, 163.70, 174.69, 185.77, 125.35, 66.43, 33.21, 0.22)
LAI <- c(0.0, 0.0, 0.2, 0.2, 0.2, 0.3, 1.3, 1.9, 1.2, 0.5, 0.3, 0.0)
P <- c(38.0, 42.0, 83.0, 73.0, 96.0, 46.0, 115.0, 173.0, 70.0, 85.0, 58.0, 72.0)
m4.ET.type1(Rn, LAI = LAI, P = P, land.type = "shrub")
}
\references{
Fang, Y., Sun, G., Caldwell, P., McNulty, S. G., Noormets, A., Domec, J. C., King, J., Zhang, Z., Zhang, X., Lin, G., Zhou, G., Xiao, J., Chen, J., and Zhou, G. (2016). Monthly land cover-specific evapotranspiration models derived from global eddy flux measurements and remote sensing data. Ecohydrology, 9(2), 248–266.
Chen, J. (2021). Biophysical Models and Applications in Ecosystem Analysis.
Higher Education Press. 104pp
}
