% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_evapotranspiration.R
\name{m4.PET.BlaneyCriddle}
\alias{m4.PET.BlaneyCriddle}
\title{Calculate PET using Blaney-Criddle model}
\usage{
m4.PET.BlaneyCriddle(ta, dhour)
}
\arguments{
\item{ta}{daily mean temperature (◦C). A vector.}

\item{dhour}{the total daytime hours for the period used. A vector.}
}
\value{
vector of potential water use (PET) for a reference crop (mm).
}
\description{
Calculates daily potential evapotranspiration (mm) based on Blaney-Criddle model.
It is used to estimate ET for the western USA. Calculated PET is the potential
water use for a reference crop.
}
\examples{
m4.PET.BlaneyCriddle(21.17, 12)
}
\references{
Doorenbos, J., and Pruitt, W. O. (1977). Guidelines for Predicting Crop Water Requirements. FAO Irrigation and Drainage Paper, 24 (revised). Rome.\cr
Chen, J. (2021). Biophysical Models and Applications in Ecosystem Analysis.
Higher Education Press. 100pp
}
