% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_evapotranspiration.R
\name{m5.PET.Turc}
\alias{m5.PET.Turc}
\title{Calculate PET using Turc model}
\usage{
m5.PET.Turc(ta, h, Rs)
}
\arguments{
\item{ta}{daily mean air temperature (◦C)}

\item{h}{relative humidity (\%). Value from 0-100. A vector.}

\item{Rs}{daily solar radiation (MJ m-2 d-1). A vector.}
}
\value{
vector of potential water use (PET) for a reference crop (mm).
}
\description{
Calculates daily evapotranspiration (mm d-1) based on Turc model.
}
\examples{
ta <- c(-3.05, -1.53, 5.31, 8.01, 15.99, 21.17, 23.17, 23.65,
19.66, 13.25, 7.61, -2.39)
h <- c(73, 74, 68, 63, 58, 60, 71, 75, 74, 77, 74, 81)
Rs <- c(-0.87, 27.20, 78.73, 95.00, 149.09, 163.70, 174.69,
185.77, 125.35, 66.43, 33.21, 0.22) # W m-2
Rs <- Rs * (1e-6*24*3600) #convert to MJ m-2 d-1
m5.PET.Turc(ta, h, Rs)
}
\references{
Doorenbos, J., and Pruitt, W. O. (1977). Guidelines for Predicting Crop Water Requirements. FAO Irrigation and Drainage Paper, 24 (revised). Rome.\cr
Chen, J. (2021). Biophysical Models and Applications in Ecosystem Analysis.
Higher Education Press. 100pp
}
