% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_evapotranspiration.R
\name{m6.PET.PriestleyTaylor}
\alias{m6.PET.PriestleyTaylor}
\title{Calculate PET using Priestley-Taylor model}
\usage{
m6.PET.PriestleyTaylor(ta, EL, G, Rn, alpha = 1.26)
}
\arguments{
\item{ta}{daily mean air temperature (◦C). A vector.}

\item{EL}{elevation (m). A vector.}

\item{G}{heat flux density to the ground (MJ m−2 d −1). A vector.}

\item{Rn}{vector of net radiation (MJ m-2). A vector.}

\item{alpha}{calibration constant, \code{alpha} = 1.26 for wet or humid conditions.}
}
\value{
vector of daily potential evapotranspiration (mm d-1)
}
\description{
Calculates daily potential evapotranspiration (mm d-1) based on Priestley-Taylor model.
}
\examples{
ta <- c(-3.05, -1.53, 5.31, 8.01, 15.99, 21.17, 23.17, 23.65,
19.66, 13.25, 7.61, -2.39)
Rn <- c(-0.87, 27.20, 78.73, 95.00, 149.09, 163.70, 174.69,
185.77, 125.35, 66.43, 33.21, 0.22) # W m-2
Rn <- Rn * (1e-6*24*3600) #convert to MJ m-2 d-1
G <- -5.52 * (1e-6*24*3600) #convert to MJ m-2 d-1
m6.PET.PriestleyTaylor(ta, 0, G, Rn)
}
\references{
Priestley, C. H. B., and Taylor, R. J. (1972). On the assessment of surface heat flux and evaporation using large-scale parameters. Monthly Weather Review, 100(2), 81–92.\cr
Chen, J. (2021). Biophysical Models and Applications in Ecosystem Analysis.
Higher Education Press. 101pp
}
