% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_evapotranspiration.R
\name{m7.PET.Makkink}
\alias{m7.PET.Makkink}
\title{Calculate PET using Makkink model}
\usage{
m7.PET.Makkink(ta, EL, Rs)
}
\arguments{
\item{ta}{daily mean air temperature (◦C)}

\item{EL}{elevation (m)}

\item{Rs}{solar radiation (MJ m-2 d-1)}
}
\value{
vector of daily potential evapotranspiration (mm d-1)
}
\description{
Calculates daily potential evapotranspiration (mm d-1) based on Makkink model.
}
\examples{
ta <- c(-3.05, -1.53, 5.31, 8.01, 15.99, 21.17, 23.17, 23.65,
19.66, 13.25, 7.61, -2.39)
Rs <- c(-0.87, 27.20, 78.73, 95.00, 149.09, 163.70, 174.69,
185.77, 125.35, 66.43, 33.21, 0.22) # W m-2
Rs <- Rs * (1e-6*24*3600) #convert to MJ m-2 d-1
m7.PET.Makkink(ta, 0, Rs)
}
\references{
Makkink, G. F. (1957). Testing the Penman formula by means of lysimeters. Journal of the Institution of Water Engineerrs, 11, 277–288.
Chen, J. (2021). Biophysical Models and Applications in Ecosystem Analysis.
Higher Education Press. 102pp
}
