% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p1_tah}
\alias{p1_tah}
\title{Temperature And Relative Humidity Data}
\format{
A data frame with 48 rows and 3 variables:
\describe{
\item{hour}{hour}
\item{Ta}{temperature, in celsius}
\item{h}{relative humidity, in percentage}
}
}
\usage{
p1_tah
}
\description{
A dataset containing the half-hour temperature and relative humidity in a day.
}
\references{
Chen, J. (2021). Biophysical Models and Applications in Ecosystem Analysis.
Higher Education Press.
}
\keyword{datasets}
