% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_photosynthesis.R
\name{photo.ballberry.fit}
\alias{photo.ballberry.fit}
\title{Fit Ball-Berry Model}
\usage{
photo.ballberry.fit(
  data,
  model = "bb",
  g0 = seq(min(data$gs)/2, mean(data$gs), 0.001)
)
}
\arguments{
\item{data}{a \code{data.frame} with several columns. It contains required data
to compute stomatal conductance for different models. The column names must be:\cr
"\code{gs}": for stomatal conductance (µmol m−2 s-1).\cr
"\code{An}": for photosynthesis rate (µmol m−2 s-1)\cr
"\code{cs}": for CO2 concentration of leaf surface (µmol mol-1)\cr
"\code{hs}": for relative humidity at the leaf surface (a value between 0-1,
rather than percentage)\cr
"\code{VPD}": vapor pressure deficit (kPa).}

\item{model}{type of model used.\cr
"\code{bb}": the original Ball-Berry model.
"\code{launing}": model proposed by Leuning in 1990 by adding parameter \code{tau}
into the original Ball-Berry model.
"\code{vpd}": model proposed by Leuning in 1995. VPD, instead of relative humidity,
is used in the model.}

\item{g0}{potential g0 for linear regression. A vector. Only used when \code{model=="launing"}.}
}
\value{
a list of four objects: \cr\cr
\code{$formula}: formula of the model.\cr
\code{$coef}: fitted coefficients of the model.\cr
\code{$r.squared}: R square of the model.\cr
\code{$adj.r.squared}: adjusted R square of the model.\cr
}
\description{
\code{photo.ballberry.fit} is used to fit Ball-Berry Model for calculating
stomatal conductance (µmol m−2 s-1). It uses linear regression after
linearization of the Ball-Berry model.
}
\examples{
cs <- seq(100, 200, 5)
data <- data.frame(cs = cs, An = rep(15, length(cs)), hs = rep(0.6, length(cs)))
data$gs <- photo.ballberry(data, g0 = 0.6, g1 = 3, tau = 20, model = "leuning") + jitter(0.01)
photo.ballberry.fit(data, model = "leuning")
}
\references{
Ball, J. T., Woodrow, I. E., and Berry, J. A. (1987). A model predicting stomatal conductance and its contribution to the control of photosynthesis under different environmental conditions. In Progress in Photosynthesis Research (pp. 221–224). Springer, Dordrecht. http://doi.org/10.1007/978-94-017-0519-6-48.\cr
Leuning, R. (1990). Modelling stomatal behaviour and photosynthesis of Eucalyptus grandis. Functional Plant Biology, 17(2), 159–175.\cr
Leuning, R. (1995). A critical appraisal of a combined stomatal‐photosynthesis model for C3 plants. Plant, Cell & Environment, 18(4), 339–355.
}
