% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_photosynthesis.R
\name{photo.ballberry}
\alias{photo.ballberry}
\title{Calculate Stomatal Conductance Based On Ball-Berry Model}
\usage{
photo.ballberry(data, g0, g1, tau = NULL, VPD_0 = NULL, model = "bb")
}
\arguments{
\item{data}{a \code{data.frame} with several columns. It contains required data
to compute stomatal conductance for different models. The column names must be:\cr
"\code{An}": for photosynthesis rate (µmol m−2 s-1)\cr
"\code{cs}": for CO2 concentration of leaf surface (µmol mol-1)\cr
"\code{hs}": for relative humidity at the leaf surface (a value between 0-1,
rather than percentage)\cr
"\code{VPD}": vapor pressure deficit (kPa).}

\item{g0}{the minimum conductance value (mol m−2 s−1). Parameter.}

\item{g1}{parameter.}

\item{tau}{parameter.}

\item{VPD_0}{the value of VPD at which stomatal conductance becomes zero (kPa). Parameter.}

\item{model}{type of model used.\cr
"\code{bb}": the original Ball-Berry model.
"\code{launing}": model proposed by Leuning in 1990 by adding parameter \code{tau}
into the original Ball-Berry model.
"\code{c3}": model proposed by Leuning in 1995. A VPD term is used for C3 plants.
"\code{vpd}": model proposed by Leuning in 1995. VPD, instead of relative humidity,
is used in the model.}
}
\value{
vector of stomatal conductance (µmol m−2 s-1)
}
\description{
\code{photo.ballberry} uses Ball-Berry model to calculate stomatal conductance
(µmol m−2 s-1) based on a set of variables.
}
\examples{
cs <- seq(100, 200, 5)
data <- data.frame(cs = cs, An = rep(15, length(cs)), hs = rep(0.6, length(cs)))
photo.ballberry(data, 0.6, 3)
}
\references{
Ball, J. T., Woodrow, I. E., and Berry, J. A. (1987). A model predicting stomatal conductance and its contribution to the control of photosynthesis under different environmental conditions. In Progress in Photosynthesis Research (pp. 221–224). Springer, Dordrecht. http://doi.org/10.1007/978-94-017-0519-6-48.\cr
Leuning, R. (1990). Modelling stomatal behaviour and photosynthesis of Eucalyptus grandis. Functional Plant Biology, 17(2), 159–175.\cr
Leuning, R. (1995). A critical appraisal of a combined stomatal‐photosynthesis model for C3 plants. Plant, Cell & Environment, 18(4), 339–355.
}
