% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_photosynthesis.R
\name{photo.farquhar}
\alias{photo.farquhar}
\title{Calculate Net Leaf CO2 Assimilation Based On Farquhar Model}
\usage{
photo.farquhar(ci, Vmax, Jmax, Q10, Tp = Inf, oi = 2.09e+08, Rd = 0)
}
\arguments{
\item{ci}{the intercellular CO2 concentration (µmol mol−1).}

\item{Vmax}{the maximum activity of Rubisco (µmol m−2 s−1).}

\item{Jmax}{the electron transport rate (µmol m−2 s−1) which can vary with absorbed
photosynthetically active radiation (aPAR).}

\item{Q10}{Q10 value of leaf respiration modeled with leaf temperature.}

\item{Tp}{the triose phosphate utilization rate (µmol m−2).}

\item{oi}{the oxygen (O2) concentration in the atmosphere (209 mol mol−1).}

\item{Rd}{respiration rate (μmol m-2 s-1).}
}
\value{
a vector of net leaf CO2 assimilation rate (An, μmol m-2 s-1).
}
\description{
calculates net leaf CO2 assimilation rate (An, μmol m-2 s-1) based on Farquhar's model.
}
\examples{
photo.farquhar(seq(10000,100000,1000), 20, 200, 3.18) ##NOTE: not validated by datasets
}
\references{
Farquhar, G. D., von Caemmerer, S. V., and Berry, J. A. (1980). A biochemical model
of photosynthetic CO2 assimilation in leaves of C3 species. Planta, 149(1), 78–90.
}
