% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_photosynthesis.R
\name{photo.landsberg.fit}
\alias{photo.landsberg.fit}
\title{Fit Landsberg Model For Photosynthesis}
\usage{
photo.landsberg.fit(PAR, Pn, Pm = seq(max(Pn)/2, max(Pn) * 2, 0.01))
}
\arguments{
\item{PAR}{photosynthetically active radiation (μmol m-2 s-1).}

\item{Pn}{measured photosynthesis rate (μmol m-2 s-1).}

\item{Pm}{potential range of the maximum photosynthetic capacity of a leaf or an
ecosystem (μmol m-2 s-1). This could be a vector, and each of its elements will
be used to fit Landsberg model, although only the model with maximal
R-square will be chosen.}
}
\value{
a list of four objects: \cr\cr
\code{$formula}: formula of the model.\cr
\code{$coef}: fitted coefficients of the model.\cr
\code{$r.squared}: R square of the model.\cr
\code{$adj.r.squared}: adjusted R square of the model.\cr
}
\description{
\code{photo.landsberg.fit} is used to fit Landsberg photosynthesis model It
uses linear regression after linearization of the Landsberg equation.
}
\examples{
PAR <- seq(200,300,10)
Pn <- photo.landsberg(PAR, alpha = 0.008, Pm = 10, Icomp = 200) + jitter(0.01) # add small noises
photo.landsberg.fit(PAR, Pn)
}
