% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_photosynthesis.R
\name{photo.landsberg}
\alias{photo.landsberg}
\title{Calculate Photosynthesis Rate Based On Landsberg Model}
\usage{
photo.landsberg(PAR, alpha = 0.008, Pm = 10, Icomp = 200)
}
\arguments{
\item{PAR}{photosynthetically active radiation (μmol m-2 s-1).}

\item{alpha}{the slope of the change in photosynthesis rate with PAR.}

\item{Pm}{the maximum photosynthetic capacity of a leaf or an ecosystem (μmol m-2 s-1).}

\item{Icomp}{the light compensation point (μmol m-2 s-1) at which the photosynthesis rate is zero.}
}
\value{
a vector of photosynthesis rate (Pn, μmol m-2 s-1).
}
\description{
calculates the photosynthesis rate (Pn, μmol m-2 s-1) based on Landsberg model.
}
\examples{
photo.landsberg(seq(200,300,10))
}
\references{
Landsberg, J. J. (1977). Some useful equations for biological studies. Experimental Agriculture,
13(3), 273–286.
}
