% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_photosynthesis.R
\name{photo.michment.fit}
\alias{photo.michment.fit}
\title{Fit Michaelis-Menten Models For Photosynthesis}
\usage{
photo.michment.fit(PAR, Pn, Rd = seq(0, max(Pn), 0.01))
}
\arguments{
\item{PAR}{photosynthetically active radiation (μmol m-2 s-1).}

\item{Pn}{measured photosynthesis rate (μmol m-2 s-1).}

\item{Rd}{potential range of respiration rate (μmol m-2 s-1). This could be a vector, and each
of its elements will be used to fit Michaelis-Menten model, although only the model with maximal
R-square will be chosen.}
}
\value{
a list of four objects: \cr\cr
\code{$formula}: formula of the model.\cr
\code{$coef}: fitted coefficients of the model.\cr
\code{$r.squared}: R square of the model.\cr
\code{$adj.r.squared}: adjusted R square of the model.\cr
}
\description{
\code{photo.michment.fit} is used to fit Michaelis-Menten photosynthesis models. It
uses linear regression after linearization of the Michaelis-Menten equation.
}
\examples{
PAR <- seq(10,100,10)
Pn <- photo.michment(PAR, alpha = 0.05, Pm=10, Rd=0) + jitter(0.01) # add small noises
photo.michment.fit(PAR, Pn)
}
