% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_photosynthesis.R
\name{photo.michment}
\alias{photo.michment}
\title{Calculate Photosynthesis Rate Based On Michaelis-Menten Model}
\usage{
photo.michment(
  PAR,
  alpha = 0.05,
  Pm = 10,
  Rd = 0,
  beta = NULL,
  method = c("michaelis", "landsand", "peat")
)
}
\arguments{
\item{PAR}{photosynthetically active radiation (μmol m-2 s-1).}

\item{alpha}{the photochemical efficiency of photosynthesis at low light.}

\item{Pm}{the maximum photosynthetic capacity of a leaf or an ecosystem (μmol m-2 s-1).}

\item{Rd}{respiration rate (μmol m-2 s-1).}

\item{beta}{additional shape factor.}

\item{method}{type of calculating methods. Option "\code{mechaelis}" adopted the
original Michaelis-Menten Model without using parameter \code{beta}. Option "\code{landsand}
adopted the non-rectangular hyperbolic model proposed by Landsberg and Sands (2011). Option "\code{peat}"
used the model applied by Peat (1970), another form of non-rectangular hyperbolic model.}
}
\value{
a vector of photosynthesis rate (Pn, μmol m-2 s-1).
}
\description{
calculates the photosynthesis rate (Pn, μmol m-2 s-1) based on Michaelis-Menten model.
}
\examples{
## Michaelis-Menten model
photo.michment(seq(10,100,10))

## Non-rectangular hyperbolic model
photo.michment(seq(10,100,10), method = "landsand", beta = -2)
photo.michment(seq(10,100,10), method = "peat", beta = -2)
}
\references{
Michaelis, L., and Menten, M. L. (1913). Die Kinetik der Invertinwirkung. Biochem Z, 49: 333–369.\cr
Landsberg, J. J., and Sands, P. (2011). Physiological Ecology of Forest Production:
Principles, Processes and Models (Vol. 4). London: Elsevier/Academic Press. 352pp.\cr
}
