% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_weather.R
\name{ra}
\alias{ra}
\title{Calculate Extraterrestrial Radiation}
\usage{
ra(lati, doy)
}
\arguments{
\item{lati}{latitude of the site in degree.}

\item{doy}{day of the year. It can be calculated by the function "\code{DOY}".}
}
\value{
vector of net solar radiation (MJ m-2 d-1).
}
\description{
calculates daily extraterrestrial radiation (MJ m-2 d-1).
}
\examples{
doy <- DOY(2022,9,3)
lati <- -20
ra(lati, doy)
}
\references{
Allen, R. G., Pereira, L. S., Raes, D. and Smith, M. (1998). Crop Evapotranspiration – Guidelines for Computing Crop Water Requirements, FAO Irrigation and Drainge Paper 56, FAO, 1998, ISBN 92-5-104219-5.
}
