% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_respiration.R
\name{respir.fit}
\alias{respir.fit}
\title{Models for Ecosystem Respiration}
\usage{
respir.fit(
  R,
  ta = NULL,
  m = NULL,
  doy = NULL,
  method = "t1.log-linear",
  n = 2,
  Q10st = 5
)
}
\arguments{
\item{R}{vector of respiration (μmol CO2 m-2 s-1).}

\item{ta}{vector of temperature (°C)}

\item{m}{vector of soil moisture (\%)}

\item{doy}{vector of day of the year}

\item{method}{type of models used. Valid options are: \code{"t1.log-linear"},
\code{"t2.linear"},  \code{"t3.polynomial"},  \code{"t4.arrhenius"},
\code{"t5.logistic"},  \code{"t6.gamma"}, \code{"tm1.DeForest"},
\code{"tm2.Xu"}, \code{"tm3.Concilio"}, \code{"tmd.Xu"}, \code{"d.fourier"}.}

\item{n}{order of polynomial model. Only used when \code{method == "t3.polynomial"}.}

\item{Q10st}{start temperature to calculate Q10 value.}
}
\value{
a list of four or five objects: \cr\cr
\code{$formula}: formula of the model.\cr
\code{$coef}: fitted coefficients of the model.\cr
\code{$r.squared}: R square of the model.\cr
\code{$adj.r.squared}: adjusted R square of the model.\cr
\code{$Q10}: Q10 value calculated from the model.
}
\description{
\code{respir.fit} is used to fit ecosystem respiration models.\cr
}
\examples{
##1 log-linear model
ta <- c(-1.64,-1.59,-1.57,-1.57,-1.51,-1.50,-1.50,-1.50,
-1.50,-1.48,-1.47,-1.42,-1.41,-1.41,-1.41,-1.39,-1.39)
R <- c(0.491,0.493,0.495,0.495,0.498,0.499,0.499,0.499,
0.499,0.500,0.500,0.503,0.504,0.504,0.504,0.505,0.505)
respir.fit(R, ta)

##2 logistic model
ta <- 0:40
R <- 5.8164/(1+exp(3.8513-0.3105*ta))
respir.fit(R, ta, method = "t5.logistic")

##3 lon then linear
tm <- expand.grid(20:40, seq(50,80,5))
ta <- tm[,1]
m <- tm[,2]
R <- 4.0896 * exp(0.0438 * ta) + 0.2178 * m -8.3419
respir.fit(R, ta, m, method = "tm1.DeForest")

##4 log then parabola
R <- 66.8272 * exp(0.1004 * ta) * 8.95373781e-6 * (m + 17.1191)^2
respir.fit(R, ta, m, method = "tm2.Xu")

##5 Concilio
R <- 319.9639 * exp(0.0310 * ta) * exp(-0.0027 * m) * 0.00003 * ta * m
respir.fit(R, ta, m, method = "tm3.Concilio")

##6 Xu
tmd <- expand.grid(5:15, seq(50,80,5), seq(100,300,25))
ta <- tmd[,1]
m <- tmd[,2]
doy <- tmd[,3]
R <- 99.29090 * exp(0.04391 * ta) * 7.37849575e-6 *
(m + 56.15651)^2 - 4.96519958e-7 * (doy + 2721.70967)^2
respir.fit(R, ta, m, doy, method = "tmd.Xu")
}
\references{
Wofsy, S. C., Goulden, M. L., Munger, J. W., Fan, S. M., Bakwin, P. S., Daube, B. C., Bassow, S. L., and Bazzaz, F. A. (1993). Net exchange of CO2 in a mid-latitude forest. Science, 260(5112), 1314–1317.\cr
Yu, X., Zha, T., Pang, Z., Wu, B., Wang, X., Chen, G., Li, C., Cao, J., Jia, G., Li, X., and Wu, H. (2011). Response of soil respiration to soil temperature and moisture in a 50-year-old oriental arborvitae plantation in China. PloS One, 6(12), e28397.\cr
Görres, C. M., Kammann, C., and Ceulemans, R. (2016). Automation of soil flux chamber measurements: Potentials and pitfalls. Biogeosciences, 13(6), 1949–1966.\cr
Barr, A. G., Griffis, T. J., Black, T. A., Lee, X., Staebler, R. M., Fuentes, J. D., Chen, Z., and Morgenstern, K. (2002). Comparing the carbon budgets of boreal and temperate deciduous forest stands. Canadian Journal of Forest Research, 32(5), 813–822.\cr
Khomik, M., Arain, M. A., Liaw, K. L., and McCaughey, J. H. (2009). Debut of a flexible model for simulating soil respiration–soil temperature relationship: Gamma model. Journal of Geophysical Research: Biogeosciences, 114(G3), doi: 10.1029/2008JG000851.\cr
DeForest, J. L., Noormets, A., McNulty, S. G., Sun, G., Tenney, G., and Chen, J. (2006) Phenophases alter the soil respiration-temperature relationship in an oak-dominated forest. International Journal of Biometeorology, 51(2), 135–144.\cr
Martin, J. G., Bolstad, P. V., Ryu, S. R., and Chen, J. Q.(2009). Modeling soil respiration based on carbon, nitrogen, and root mass across diverse Great Lake forests. Agricultural and Forest Meteorology, 149(10), 1722–172.\cr
Xu, M., and Qi, Y. (2001). Soil-surface CO2 efflux and its spatial and temporal variations in a young ponderosa pine plantation in northern California. Global Change Biology, 7(6), 667–677.\cr
Concilio, A., Ma, S., Li, Q., LeMoine, J., Chen, J., North, M., Moorhead, D., and Jensen, R. (2005). Soil respiration response to prescribed burning and thinning in mixed-conifer and hardwood forests. Canadian Journal of Forest Research, 35(7), 1581–1591.\cr
Davidson, E. A., Janssens, I. A., and Luo, Y. (2006). On the variability of respiration in terrestrial ecosystems: Moving beyond Q10. Global Change Biology, 12(2), 154–164.\cr
}
