% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_weather.R
\name{rn.m1}
\alias{rn.m1}
\title{Calculate Net Solar Radiation (Method 1)}
\usage{
rn.m1(ta, tmax, tmin, Rs, Ra)
}
\arguments{
\item{ta}{mean air temperature (◦C). A vector.}

\item{tmax}{maximum air temperature (◦C). A vector.}

\item{tmin}{minimum air temperature (◦C). A vector.}

\item{Rs}{solar radiation (MJ m-2 d-1).}

\item{Ra}{extraterrestrial solar radiation (MJ m-2 d-1).}
}
\value{
vector of net solar radiation (MJ m-2 d-1).
}
\description{
calculates net solar radiation values (MJ m-2 d-1) based on (incoming) and extraterrestrial solar radiation.
}
\examples{
lati <- -22.9
doy <- DOY(2022, 5, 15)
ah <- 7.1
Rs <- rs(lati, doy, ah)
Ra <- ra(lati, doy)
rn.m1(25, 30, 20, Rs, 120)
}
\references{
Chen, J. (2021). Biophysical Models and Applications in Ecosystem Analysis.
Higher Education Press. 102pp
}
