% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_weather.R
\name{rn.m2}
\alias{rn.m2}
\title{Calculate Net Radiation (Method 2)}
\usage{
rn.m2(Rns, Rnl)
}
\arguments{
\item{Rns}{net shortwave radiation (MJ m-2 day-1). It can be calculated by the function "\code{rns}".}

\item{Rnl}{net longwave radiation (MJ m-2 day-1). It can be calculated by the function "\code{rnl}".}
}
\value{
values of net radiation (MJ m-2 day-1)
}
\description{
calculates net solar radiation (MJ m-2 day-1) based on net shortwave and longwave radiation.
}
\examples{
lati <- -22.7
doy <- DOY(2022, 5, 15)
ah <- 7.1
Rs <- rs(lati, doy, ah)
Rso <- rso(lati, doy)
Rns <- rns(Rs)
Rnl <- rnl(Rs, Rso, 2.1, 25.1, 19.1)
rn.m2(Rns, Rnl)
}
\references{
Allen, R. G., Pereira, L. S., Raes, D. and Smith, M. (1998). Crop Evapotranspiration – Guidelines for Computing Crop Water Requirements, FAO Irrigation and Drainge Paper 56, FAO, 1998, ISBN 92-5-104219-5.
}
