% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_weather.R
\name{rnl}
\alias{rnl}
\title{Calculate Net Longwave Radiation}
\usage{
rnl(Rs, Rso, ea, tmax, tmin)
}
\arguments{
\item{Rs}{(incoming) solar radiation (MJ m-2 d-1). It can be calculated by the function "\code{rs}" if it is not measured.}

\item{Rso}{short wave radiation on a clear-sky day (MJ m-2 d-1). It can be calculated by the function "\code{rso}".}

\item{ea}{actual vapor pressure (kPa). It can be calculated by the function "\code{actualvp}".}

\item{tmax}{maximum absolute temperature during a day (◦C).}

\item{tmin}{minimum absolute temperature during a day (◦C).}
}
\value{
value(s) of net longwave radiation (MJ m-2 d-1).
}
\description{
calculates net longwave radiation (MJ m-2 d-1).
}
\examples{
lati <- -22.7
doy <- DOY(2022, 5, 15)
ah <- 7.1
Rs <- rs(lati, doy, ah)
Rso <- rso(lati, doy)
rnl(Rs, Rso, 2.1, 25.1, 19.1)
}
\references{
Allen, R. G., Pereira, L. S., Raes, D. and Smith, M. (1998). Crop Evapotranspiration – Guidelines for Computing Crop Water Requirements, FAO Irrigation and Drainge Paper 56, FAO, 1998, ISBN 92-5-104219-5.
}
