% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_weather.R
\name{rns}
\alias{rns}
\title{Calculate Net Shortwave Radiation}
\usage{
rns(Rs, albedo = 0.23)
}
\arguments{
\item{Rs}{(incoming) solar radiation (MJ m-2 d-1). It can be calculated by the function "\code{rs}" if it is not measured.}

\item{albedo}{albedo or canopy reflection coefficient, which is 0.23 for the hypothetical grass reference crop.}
}
\value{
values of net shortwave radiatin (MJ m-2 day-1).
}
\description{
calculates net shortwave radiation (MJ m-2 day-1)
}
\examples{
Rs <- rs(-22.9, DOY(2022, 5, 15), 7.1)
rns(Rs)
}
\references{
Allen, R. G., Pereira, L. S., Raes, D. and Smith, M. (1998). Crop Evapotranspiration – Guidelines for Computing Crop Water Requirements, FAO Irrigation and Drainge Paper 56, FAO, 1998, ISBN 92-5-104219-5.
}
