% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_weather.R
\name{rs}
\alias{rs}
\title{Calculate Solar Radiation}
\usage{
rs(lati, doy, ah, as = 0.25, bs = 0.5)
}
\arguments{
\item{lati}{latitude of the site in degree.}

\item{doy}{day of the year. It can be calculated by the function "\code{DOY}".}

\item{ah}{actual duration of sunshine in a day (h).}

\item{as}{regression constant, expressing the fraction of extraterrestrial radiation reaching the earth on overcast days. Default 0.25.}

\item{bs}{fraction of extraterrestrial radiation reaching the earth on clear days. Default 0.5.}
}
\value{
value(s) of incoming solar radiation (or shortwave radiation, MJ m-2 d-1).
}
\description{
calculates (incoming) solar radiation when it is not measured.
}
\examples{
rs(-22.9, DOY(2022, 5, 15), 7.1)
}
\references{
Allen, R. G., Pereira, L. S., Raes, D. and Smith, M. (1998). Crop Evapotranspiration – Guidelines for Computing Crop Water Requirements, FAO Irrigation and Drainge Paper 56, FAO, 1998, ISBN 92-5-104219-5.
}
