% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_weather.R
\name{rso}
\alias{rso}
\title{Calculate Short Wave Radiation}
\usage{
rso(lati, doy, tau = 0.75, EL = 0, method = "tau")
}
\arguments{
\item{lati}{latitude of the site in degree.}

\item{doy}{day of the year. It can be calculated by the function "\code{DOY}".}

\item{tau}{clear sky transmissivity. Default 0.75.}

\item{EL}{station elevation (m). Default 0.}

\item{method}{method to calculate short wave radiation. Valid options are:\cr
"\code{tau}": use clear sky transmissivity.\cr
"\code{EL}": use elevation.\cr}
}
\value{
value(s) of short wave radiation (MJ m-2 d-1).
}
\description{
calculates short wave radiation on a clear-sky day (MJ m-2 d-1).
}
\examples{
rso(-22.9, DOY(2022, 5, 15))
}
\references{
Allen, R. G., Pereira, L. S., Raes, D. and Smith, M. (1998). Crop Evapotranspiration – Guidelines for Computing Crop Water Requirements, FAO Irrigation and Drainge Paper 56, FAO, 1998, ISBN 92-5-104219-5.
}
