% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_weather.R
\name{satvd}
\alias{satvd}
\title{Calculate Saturation Vapor Density}
\usage{
satvd(Ta, ...)
}
\arguments{
\item{Ta}{vector of air temperature (◦C).}

\item{...}{additional parameters accepted by function \code{satvp}.}
}
\value{
vector of saturation vapor density (kg m-3).
}
\description{
calculates saturation vapor density (kg m-3). Only valid for temperature above
0◦C.
}
\examples{
data("p1_tah")
temperature <- p1_tah$Ta
satvd(temperature)
}
