% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_weather.R
\name{satvp}
\alias{satvp}
\title{Calculate Saturation Vapor Pressure}
\usage{
satvp(Ta, method = c("Tetens", "Lowry"))
}
\arguments{
\item{Ta}{vector of air temperature (◦C).}

\item{method}{type of estimator, either \code{"Tetens"} for model developed by
Monteith and Unsworth (2013) or \code{"Lowry"} for model developed by
Fritschen and Gay (1979).}
}
\value{
vector of saturation vapor pressure (kPa).
}
\description{
calculates saturation vapor pressure (kPa). Only valid for temperature above
0◦C.
}
\examples{
data("p1_tah")
temperature <- p1_tah$Ta

satvp(temperature) ##1 Tetens method
satvp(temperature, method="Lowry") ##2 Lowry method
}
\references{
Monteith, J., and Unsworth, M. (2013). Principles of Environmental Physics:
Plants, Animals, and the Atmosphere. Academic Press. 423pp.\cr
Lowry, W. P. (2013). Weather and Life: An Introduction to Biometeorology.
Elsevier. 305pp.Monteith, J., and Unsworth, M. (2013). Principles of
Environmental Physics: Plants, Animals, and the Atmosphere. Academic Press. 423pp\cr
Fritschen, L. J., and Gay, L. W. (1979). Environmental Instrumentation.
Springer Science & Business Media. 216pp.
}
