% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_gwp.R
\name{transmittance}
\alias{transmittance}
\title{Calculate Upwelling Transmittance Value}
\usage{
transmittance(Sw_down, doy, zenith)
}
\arguments{
\item{Sw_down}{the total incident sunlight (W m-2). A vector or a scalar.}

\item{doy}{day of the year. It can be calculated by function "\code{DOY}". A vector or a scalar.}

\item{zenith}{zenith angle. It can be calculated by function "\code{zangle}". A vector or a scalar.}
}
\value{
upwelling transmittance value(s).
}
\description{
calculates upwelling transmittance value.
}
\examples{
doy <- DOY(2018, 7, 27)
zenith <- zangle(2018, 7, 27, 85.44675833, 42.47670833, 9.4)[[2]]
transmittance(691.55, doy, zenith)
}
