% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_weather.R
\name{wetbulbT}
\alias{wetbulbT}
\title{Calculate Wet-bulb Temperature}
\usage{
wetbulbT(Ta, h, ..., Cp = 29.3, lambda = 40660)
}
\arguments{
\item{Ta}{vector of air temperature (◦C).}

\item{h}{vector of relative humidity (\%).}

\item{...}{additional parameters accepted by function \code{satvp}.}

\item{Cp}{specific heat of air (29.3 J mol-1 ◦C-1).}

\item{lambda}{latent heat vaporization of water (40.660 kJ mol-1)}
}
\value{
vector of dew point temperature (◦C).
}
\description{
calculates Wet-bulb temperature (◦C). Only valid for temperature above
0◦C.
}
\examples{
data("p1_tah")
temperature <- p1_tah$Ta
humidity <- p1_tah$h
wetbulbT(temperature, humidity)
}
\references{
Fritschen, L. J., and Gay, L. W. (1979). Environmental Instrumentation.
Springer Science & Business Media. 216pp
}
