% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_weather.R
\name{ws}
\alias{ws}
\title{Calculate Vertical Profile of Wind Speed}
\usage{
ws(
  z,
  uz = NULL,
  z0 = 0.01474926,
  d = 0.07994099,
  kappa = 0.4,
  ustar = 0.1,
  windu.dmax = 2,
  mode = "wind2"
)
}
\arguments{
\item{z}{specified height (above the ground level) of the wind speed measurement (m). A vector.}

\item{uz}{measured wind speed at height \code{z}. A vector of the same length as parameter \code{z}.}

\item{z0}{the roughness length (m) at which wind speed is near zero.}

\item{d}{zero plain displacement (m).}

\item{kappa}{the von Karmon constant with an average value of 0.35–0.43.
A value of 0.40 is often used in the literature.}

\item{ustar}{the friction velocity (m s-1) which depends on the shear stress
(kg m-1 s-2) at the boundary of the flow and air density (kg m-3).}

\item{windu.dmax}{maximal value of parameter \code{d} to be considered when fiting the
wind profile model.}

\item{mode}{three options.\cr
"\code{wind2}" convert wind speed measured at a certain
height to the wind speed at 2meters, the same as function "\code{wind2}" in package
\code{sirad}. Parameter \code{kappa}, \code{ustar}, and \code{windu.dmax} will
not be used.\cr
"\code{windz}" calculates wind speed at any height based on wind profile model.
Parameter \code{windu.dmax} will not be used.\cr
"\code{windu}" fits the wind profile model using measured wind speed data at
certain heights. Parameter \code{z0}, \code{d}, and \code{ustar} will not be used
and they will be estimated by regression. Models with zero plain displacement over
than \code{windu.dmax} will not be considered.}
}
\value{
When \code{mode = "wind2"}, it returns a vector of wind speed (m s-1) at
2 meters. When \code{mode = "windz"}, it returns a vector of vertical profile of
wind speed (m s-1) for different height (indicated by the name of the vector).
When \code{mode = "windu"}, it returns a list with four components: formula,
estimated parameters, R square, and adjusted R square value of the regression.
}
\description{
calculates vertical profile of wind speed, or fit the model of wind profile.
Note that this algorithm is valid only for the neutral atmosphere and should
not be applied for tall canopies or vegetation with large variation across
the horizontal space.
}
\examples{
## calculate wind profile
z <- seq(1,10,0.1)
wp <- windspeed(z, mode = "windz")
plot(z,wp)
## calculate wind speed at 2m based on measured speed at 9m
windspeed(z=9, uz=1.601216)

## estimate parameters based on measured wind speeds
md <- windspeed(z, wp, mode = "windu")
md$coef
}
\references{
Chen, J. (2021). Biophysical Models and Applications in Ecosystem Analysis.
Higher Education Press. 39pp
}
