% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_weather.R
\name{zangle}
\alias{zangle}
\title{Calculate Solar Zenith/Azimuth Angle}
\usage{
zangle(year, month, day, long, lat, time_series = 0:23)
}
\arguments{
\item{year}{a scalar.}

\item{month}{month of year (1-12). A scalar.}

\item{day}{day of the month (1-30 or 31). A scalar.}

\item{long}{longtitude of the site (-180 to 180 degree). A scalar.}

\item{lat}{latitude of the site (-90 to 90 degree). A scalar.}

\item{time_series}{vector of time in a day. Values within 0-24 (hour).}
}
\value{
a \code{"data.frame"} of four columns. The first column is time of a
day. The last three columns show solar zenith, azimuth, and solar elevation angles, respectively. Unit in degrees.
}
\description{
calculates solar zenith, azimuth, and elevation angle.
}
\examples{
zangle(2016, 9, 22, 85.44675833, 42.47670833)
}
\references{
Chen, J. (2021). Biophysical Models and Applications in Ecosystem Analysis.
Higher Education Press. 33pp
}
